+++ src/components/EditorRightSideBar/DisplayAdjustment/index.tsx
